UI fix to insure the right header height
This change applies to when the user removes the end date of a range.
The problem was that the old, two inputs tall size was kept, preventing
the calendar from being complety hidden.

The proposed fix uses an old and dirty hack: layout trashing.
It basically removes the end input from the layout, asks for a re-layout
by triggering both update and set size events and then reverts the
initial display value and starts the animation.

There are other ways to fix it, but since the system is event based, I
afraid it would take to much effort.