Make sure 00:00:00 is saved when time is disabled
This change makes sure that the time is always midnight (00:00:00) when
the field setting 'time' is set to 0.

This change will also simplify filtering with 'now' when time is
disabled.

It introduce the method `parseDateTime` in order to acheive it without
code duplication. This method handles the proper date format according
to the field's setting.

Fixes #159